usb-export: support deauthorized devices
Previously, usb-export failed when used with deauthorized USB devices.
This commit adds support for deauthorized USB devices, by enabling
authorization before binding usbip-host to the device, and restoring the
previous setting at cleanup.

One reason a user might deauthorize USB devices in a USB VM is in order
to prevent USB devices from automatically being bound to drivers (which
gives malicious devices a bigger attack surface). Before binding
usbip-host to the device, it needs to be authorized again, so in order
to prevent the device from being bound to another driver in the
meantime, we temporarily disable drivers_autoprobe.